refactor(web): retire the hand-rolled resource hook
`useResource` has no callers left, so it and its test go, and what the pages
still use — `LoadState`, `AsyncState`, `isTerminalResourceError` — merges with
the query adapter into one `resource-state` module.

The two remaining hand-rolled polls stay, with a comment recording why. Neither
reads a resource: Computer setup waits on a mutation's side effect against a
baseline that has to be fresh at click time, and Feishu setup drives one
authorization to completion. Nothing else reads either, so there is nothing to
share, and the lifecycles they keep are the substance of those flows.
