+++ apps/web/src/features/agents/agent-settings/agent-settings-page.tsx
@@ -4,8 +4,7 @@ import { browserApi } from "../../../api.js";
 import { queryKeys } from "../../../query/keys.js";
 import { Icon, Loader, Text } from "../../../ui/design-system.js";
 import { NotFoundPage } from "../../not-found.js";
-import { toResourceState } from "../../resource/query-state.js";
-import { AsyncState } from "../../resource/use-resource.js";
+import { AsyncState, toResourceState } from "../../resource/resource-state.js";
 import type { AgentDetailView } from "../agent-model.js";
 import { useAgentDetailView } from "../agent-queries.js";
 import { agentDetailLink, agentSettingsLink, agentSettingsSectionLink } from "../agent-routes.js";
