+++ apps/web/src/features/agents/new-agent-page.tsx
@@ -7,9 +7,8 @@ import { FeishuSetup } from "../../im/feishu-setup.js";
 import { queryKeys } from "../../query/keys.js";
 import { Button, Dialog, Text } from "../../ui/design-system.js";
 import { Page } from "../layout/page.js";
-import { toResourceState } from "../resource/query-state.js";
-import type { LoadState } from "../resource/use-resource.js";
-import { AsyncState } from "../resource/use-resource.js";
+import type { LoadState } from "../resource/resource-state.js";
+import { AsyncState, toResourceState } from "../resource/resource-state.js";
 import { useWorkspace } from "../session/session-context.js";
 import { useComputersQuery } from "./agent-queries.js";
 import { agentDetailLink } from "./agent-routes.js";
