+++ apps/web/src/routes/_authenticated/internal.onboarding-lab.tsx
@@ -2,8 +2,7 @@ import { useQuery } from "@tanstack/react-query";
 import { createFileRoute, useNavigate } from "@tanstack/react-router";
 import { browserApi } from "../../api.js";
 import { StandaloneNotFoundPage } from "../../features/not-found.js";
-import { toResourceState } from "../../features/resource/query-state.js";
-import { AsyncState } from "../../features/resource/use-resource.js";
+import { AsyncState, toResourceState } from "../../features/resource/resource-state.js";
 import { useAccount } from "../../features/session/session-context.js";
 import { OnboardingLabPage } from "../../internal/onboarding-lab-page.js";
 import { queryKeys } from "../../query/keys.js";
